Add files via upload
The purpose of the course work is to create a Windows application that emulates the movement of the base figure in such a way that its center describes a given trajectory. When moving along the trajectory, the figure simultaneously rotates around the central point, and also changes the color of the unshaded part of the figure from blue (at the beginning of the trajectory) to red (at the end of the trajectory), the shaded elements – from yellow to blue.
Also in this work, it is necessary to provide for the possibility of changing the speed of rotation and movement of the figure by the user, and the possibility of scaling the figure while maintaining its proportions and the possibility of returning to its original size, and the initial direction of the trajectory in the form of a displacement vector. These features will be implemented using graphical interface elements.
To implement the project, the C# programming language will be used, using the Visual Studio 2019 development environment.